% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_datos_fred.R
\name{base_datos_fred}
\alias{base_datos_fred}
\title{Esta funcion obtiene data diaria}
\usage{
base_datos_fred(activos, fecha.i, fecha.f)
}
\arguments{
\item{activos}{conjunto de Codigos a descargar de la base del FRED}

\item{fecha.i}{fecha inicial "c(yyyy,m,d)"}

\item{fecha.f}{fecha inicial "c(yyyy,m,d)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data diaria
}
