% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_datos_m.R
\name{base_datos_m}
\alias{base_datos_m}
\title{Esta funcion obtiene data trimestral}
\usage{
base_datos_m(activos, fecha.i, fecha.f)
}
\arguments{
\item{fecha.i}{fecha inicial "c(yyyy,q)"}

\item{fecha.f}{fecha inicial "c(yyyy,q)"}

\item{actvos}{conjunto  de Codigos a descargar de la base de BCRP}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data trimestral
}
