% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_d.R
\name{datos_d}
\alias{datos_d}
\title{Esta funcion obtiene Data diaria
##########################################
fecha.i = c(yyyy,m,d)
fecha.f = c(yyyy,m,d)
"2002-01-01/2022-04-28"}
\usage{
datos_d(codigo, fecha.i, fecha.f)
}
\arguments{
\item{codigo}{Codigo a descargar de la base de BCRP}

\item{fecha.i}{fecha inicial "c(yyyy,q)"}

\item{fecha.f}{fecha inicial "c(yyyy,q)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene Data diaria
##########################################
fecha.i = c(yyyy,m,d)
fecha.f = c(yyyy,m,d)
"2002-01-01/2022-04-28"
}
