% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_d_fred.R
\name{datos_d_fred}
\alias{datos_d_fred}
\title{Esta funcion obtiene data diaria}
\usage{
datos_d_fred(codigo, fecha.i, fecha.f)
}
\arguments{
\item{codigo}{Codigo a descargar de la base del FRED}

\item{fecha.i}{fecha inicial "c(yyyy,m,d)"}

\item{fecha.f}{fecha inicial "c(yyyy,m,d)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data diaria
}
