% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_m_fred.R
\name{datos_m_fred}
\alias{datos_m_fred}
\title{Esta funcion obtiene data mensual}
\usage{
datos_m_fred(codigo, fecha.i, fecha.f)
}
\arguments{
\item{codigo}{Codigo a descargar de la base del FRED}

\item{fecha.i}{fecha inicial "c(yyyy,q)"}

\item{fecha.f}{fecha inicial "c(yyyy,q)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data mensual
}
