% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_mensual.R
\name{datos_mensual}
\alias{datos_mensual}
\title{Esta funcion obtiene data mensual}
\usage{
datos_mensual(codigo, fecha.i, fecha.f)
}
\arguments{
\item{codigo}{Codigo a descargar de la base de BCRP}

\item{fecha.i}{fecha inicial "c(yyyy,q)"}

\item{fecha.f}{fecha inicial "c(yyyy,q)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data mensual
}
