% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos_trimestral.R
\name{datos_trimestral}
\alias{datos_trimestral}
\title{Esta funcion obtiene data trimestral}
\usage{
datos_trimestral(codigo, fecha.i, fecha.f)
}
\arguments{
\item{codigo}{Codigo a descargar de la base de BCRP}

\item{fecha.i}{fecha inicial "c(yyyy,q)"}

\item{fecha.f}{fecha inicial "c(yyyy,q)"}
}
\value{
Dataframe con datos
}
\description{
Esta funcion obtiene data trimestral
}
